Create stock-analysis
This code loads stock data from a CSV file, calculates the daily returns, and the 30-day and 60-day moving averages. It then plots the daily returns and moving averages using matplotlib. This code is a simple example of how to use Python to perform financial analysis and visualize the results. Note that you will need to have the pandas and matplotlib libraries installed to run this code.