Create simple-moving-average
This code uses the yfinance library to retrieve the historical stock data for Apple (AAPL) and stores it in a Pandas dataframe. It then calculates the simple moving average with a window of 20 days and stores it in a new column called "SMA20". Finally, it prints the last five rows of the dataframe which includes the stock data and the calculated simple moving average. This code is a simple example of how to use Python to retrieve stock data and perform financial analysis using the simple moving average.